refactor: move all types into their respective files from types.ts file closes #19
